+++ api/python/quilt3/backends/base.py
@@ -2,13 +2,6 @@ import abc
 import operator
 import time
 
-from quilt3.data_transfer import (
-    copy_file,
-    delete_url,
-    get_bytes,
-    list_url,
-    put_bytes,
-)
 from quilt3.util import PhysicalKey, QuiltException
 
 
@@ -136,12 +129,12 @@ class PackageRegistryV1(PackageRegistry):
 
     def push_manifest(self, pkg_name: str, top_hash: str, manifest_data: bytes):
         """returns: timestamp to support catalog drag-and-drop => browse"""
-        put_bytes(manifest_data, self.manifest_pk(pkg_name, top_hash))
+        self.manifest_pk(pkg_name, top_hash).put_bytes(manifest_data)
         hash_bytes = top_hash.encode()
         # TODO: use a float to string formatter instead of double casting
         timestamp_str = str(int(time.time()))
-        put_bytes(hash_bytes, self.pointer_pk(pkg_name, timestamp_str))
-        put_bytes(hash_bytes, self.pointer_latest_pk(pkg_name))
+        self.pointer_pk(pkg_name, timestamp_str).put_bytes(hash_bytes)
+        self.pointer_latest_pk(pkg_name).put_bytes(hash_bytes)
         return timestamp_str
 
     @staticmethod
@@ -153,7 +146,7 @@ class PackageRegistryV1(PackageRegistry):
             top_hash = hash_prefix
         elif 6 <= len(hash_prefix) < 64:
             matching_hashes = [self._top_hash_from_path(h) for h, _
-                               in list_url(self.manifests_package_dir(pkg_name))
+                               in self.manifests_package_dir(pkg_name).list_url()
                                if h.startswith(hash_prefix)]
             if not matching_hashes:
                 raise QuiltException("Found zero matches for %r" % hash_prefix)
@@ -171,7 +164,7 @@ class PackageRegistryV1(PackageRegistry):
     def shorten_top_hash(self, pkg_name: str, top_hash: str) -> str:
         min_shorthash_len = 7
 
-        matches = [self._top_hash_from_path(h) for h, _ in list_url(self.manifests_package_dir(pkg_name))
+        matches = [self._top_hash_from_path(h) for h, _ in self.manifests_package_dir(pkg_name).list_url()
                    if h.startswith(top_hash[:min_shorthash_len])]
         if len(matches) == 0:
             raise ValueError(f"Tophash {top_hash} was not found in registry {self.base}")
@@ -185,6 +178,8 @@ class PackageRegistryV1(PackageRegistry):
         return self.list_package_pointers(pkg_name)
 
     def delete_package_version(self, pkg_name: str, top_hash: str):
+        from quilt3.data_transfer import copy_file
+
         deleted = []
         remaining = []
         for path, pkg_hash in self.list_package_versions(pkg_name):
@@ -192,7 +187,7 @@ class PackageRegistryV1(PackageRegistry):
         if not deleted:
             raise QuiltException("No such package version exists in the given directory.")
         for path in deleted:
-            delete_url(self.pointer_pk(pkg_name, path))
+            self.pointer_pk(pkg_name, path).delete_url()
         if 'latest' in deleted and remaining:
             # Create a new "latest". Technically, we need to compare numerically,
             # but string comparisons will be fine till year 2286.
@@ -247,8 +242,8 @@ class PackageRegistryV2(PackageRegistry):
             yield str(int(dt.timestamp())), top_hash
 
     def push_manifest(self, pkg_name: str, top_hash: str, manifest_data: bytes):
-        put_bytes(manifest_data, self.manifest_pk(pkg_name, top_hash))
-        put_bytes(top_hash.encode(), self.pointer_latest_pk(pkg_name))
+        self.manifest_pk(pkg_name, top_hash).put_bytes(manifest_data)
+        self.pointer_latest_pk(pkg_name).put_bytes(top_hash.encode())
 
     @staticmethod
     def _top_hash_from_path(path: str) -> str:
@@ -259,12 +254,12 @@ class PackageRegistryV2(PackageRegistry):
     shorten_top_hash = PackageRegistryV1.shorten_top_hash
 
     def delete_package_version(self, pkg_name: str, top_hash: str):
-        delete_url(self.manifest_pk(pkg_name, top_hash))
-        if get_bytes(self.pointer_latest_pk(pkg_name)).decode() == top_hash:
-            delete_url(self.pointer_latest_pk(pkg_name))
+        self.manifest_pk(pkg_name, top_hash).delete_url()
+        if self.pointer_latest_pk(pkg_name).get_bytes().decode() == top_hash:
+            self.pointer_latest_pk(pkg_name).delete_url()
             timestamp, new_latest = max(self.list_package_versions_with_timestamps(pkg_name), default=(None, None))
             if new_latest:
-                put_bytes(new_latest.encode(), self.pointer_latest_pk(pkg_name))
+                self.pointer_latest_pk(pkg_name).put_bytes(new_latest.encode())
         for pointer, pointer_top_hash in self.list_package_pointers(pkg_name):
             if pointer_top_hash == top_hash:
-                delete_url(self.pointer_pk(pkg_name, pointer))
+                self.pointer_pk(pkg_name, pointer).delete_url()
